New CLI interface
A complete rewrite of the CLI interface. The r2 commands are not hidden
when executing in standalone. This will make the commands more
discoverable. Also, the standalone commands can be executed while
running in r2 mode. The version command was updated to include compiler
version and GoRE version used. An ASCII banner was also added.