Go 1.18 package classification fix
This pulls in the latest version of gore to fix edge-cases in the
package classification logic for Go 1.18 binaries. In addition to the
update, an additional logic was added to the mod command. If the main
module path is empty, it uses the path found in the build info
structure.